+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Assert-ObjectProperty`
+  - Added parameter `Each` to opt-in for element-by-element checking when
+    passing arrays via pipeline. Without this parameter, the command now checks
+    properties on the array object itself (e.g., `Count`, `Length`).
+
 ## [1.1.0] - 2025-10-03
 
 ### Added